+++ Client/src/components/Tournament/PlayerDataGrid.js
@@ -10,7 +10,6 @@ import DataGrid, {
     FilterRow,
     FilterPanel,
     FilterBuilderPopup,
-    Scrolling,
     Paging
 } from 'devextreme-react/data-grid';
 
@@ -41,7 +40,6 @@ export default function PlayerDataGrid({ players }) {
                     <FilterRow visible={true} />
                     <FilterPanel visible={true} />
                     <FilterBuilderPopup position={filterBuilderPopupPosition} />
-                    <Scrolling columnRenderingMode="virtual" />
                     <Paging enabled={false} />
 
                     <Column dataField="ProfileName" caption="Name" width={150} fixed={true} cellRender={cellLink} />